feat: move processing clustervariable classes to its own package
